Add an entryset-based parallel citation system.
This is overly complicated in part because we can't just put fields
directly onto the entryset (biber is hardcoded to drop them).
Unfortunately, entrysets seem to be the only entry type where biber will
automatically include referenced entries.

It would be very nice to get functionality added to biber to make this
more workable.